+++ lib/cgraph/cgraph.h
@@ -58,13 +58,6 @@ extern "C" {
 #define CGRAPH_API /* nothing */
 #endif
 
-#ifndef FALSE
-#define FALSE (0)
-#endif
-#ifndef TRUE
-#define TRUE (!FALSE)
-#endif
-
 /// @endcond
 
 /// @addtogroup cgraph_misc